--- src/camadafp.h
@@ -1,45 +1,0 @@
-/**************************************************************************
- *
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- *
- **************************************************************************/
-
-#ifndef CAMADAFP_H_
-#define CAMADAFP_H_
-
-namespace camada {
-
-enum class FPEncoding { Native, BV };
-
-enum class FPNegBehavior {
-  FlipSignBit,
-  // Follows SMT-LIB FP semantics: fp.neg leaves NaNs unchanged.
-  PreserveNaNPayload,
-};
-
-enum class RM {
-  ROUND_TO_EVEN = 0,
-  ROUND_TO_AWAY = 1,
-  ROUND_TO_PLUS_INF = 2,
-  ROUND_TO_MINUS_INF = 3,
-  ROUND_TO_ZERO = 4,
-};
-
-} // namespace camada
-
-#endif
