+++ src/yicessolver.cpp
@@ -23,8 +23,7 @@
 
 #include "yicessolver.h"
 #include "camada.h"
-#include "camadaerror.h"
-#include "camadafp.h"
+#include "camadacommon.h"
 
 #include <cstdio>
 #include <gmp.h>
